Some Changes in UI of Listing
Currently I am working on the UI of the Project. I know there are so many backend things to work on. There will one more commit for the Ui. Then UI is completed and I will commit the Backend files only. B the End of this week I will try to backend and make the UI functional instead of just fake cards